+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unrealesed]
+
+### Fixed
+- Enable emergency restore on invalid pin wallet
+
 ## [4.0.1] - 2023-04-18
 
 ### Changed